Fix an issue with dropped sync callbacks
When doing camonitor with a sync callback, if the
PV receives 2 or more updates between connect and
the first callback firing, then only the first
callback will be called. This only happens when
all_updates=True because with all_updates=False
they will be squashed into a single callback with
the latest value.

The issue is that __signal wakes up
__create_subscription when there are already
multiple pending_values, but __create_subscription
handles the first, finds it is sync, then hands
back to __signal, which is not called on the
remaining values. This fixes it by making
__create_subscription handle all the updates it
finds on the initial pass